Crate google_androidmanagement1

This documentation was generated from Android Management crate version 1.0.9+20190624, where 20190624 is the exact revision of the androidmanagement:v1 schema built by the mako code generator v1.0.9.

Everything else about the Android Management v1 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

enterprises

applications get, create, devices delete, devices get, devices issue command, devices list, devices operations cancel, devices operations delete, devices operations get, devices operations list, devices patch, enrollment tokens create, enrollment tokens delete, get, patch, policies delete, policies get, policies list, policies patch, web apps create, web apps delete, web apps get, web apps list, web apps patch and web tokens create

signup urls

create

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.enterprises().policies_list(...).doit()
let r = hub.enterprises().web_apps_delete(...).doit()
let r = hub.enterprises().policies_patch(...).doit()
let r = hub.enterprises().create(...).doit()
let r = hub.enterprises().web_apps_create(...).doit()
let r = hub.enterprises().enrollment_tokens_create(...).doit()
let r = hub.enterprises().web_apps_list(...).doit()
let r = hub.enterprises().devices_list(...).doit()
let r = hub.enterprises().policies_delete(...).doit()
let r = hub.enterprises().devices_operations_cancel(...).doit()
let r = hub.enterprises().devices_operations_delete(...).doit()
let r = hub.enterprises().devices_operations_get(...).doit()
let r = hub.enterprises().devices_operations_list(...).doit()
let r = hub.enterprises().devices_patch(...).doit()
let r = hub.enterprises().patch(...).doit()
let r = hub.enterprises().policies_get(...).doit()
let r = hub.enterprises().web_tokens_create(...).doit()
let r = hub.enterprises().web_apps_patch(...).doit()
let r = hub.enterprises().enrollment_tokens_delete(...).doit()
let r = hub.enterprises().devices_issue_command(...).doit()
let r = hub.enterprises().devices_delete(...).doit()
let r = hub.enterprises().web_apps_get(...).doit()
let r = hub.enterprises().devices_get(...).doit()
let r = hub.enterprises().get(...).doit()
let r = hub.enterprises().applications_get(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-androidmanagement1 = "*"
# This project intentionally uses an old version of Hyper. See
# https://github.com/Byron/google-apis-rs/issues/173 for more
# information.
hyper = "^0.10"
hyper-rustls = "^0.6"
serde = "^1.0"
serde_json = "^1.0"
yup-oauth2 = "^1.0"

A complete example

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_androidmanagement1 as androidmanagement1;
use androidmanagement1::Enterprise;
use androidmanagement1::{Result, Error};
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidmanagement1::AndroidManagement;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidManagement::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Enterprise::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.enterprises().create(req)
             .signup_url_name("sed")
             .project_id("et")
             .enterprise_token("dolores")
             .doit();
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

AlwaysOnVpnPackage
Configuration for an always-on VPN connection.

AndroidManagement
Central instance to access all AndroidManagement related resource activities

ApiLevelCondition
A compliance rule condition which is satisfied if the Android Framework API level on the device doesn't meet a minimum requirement. There can only be one rule with this type of condition per policy.

Application
Information about an app.

ApplicationEvent
An app-related event.

ApplicationPermission
A permission required by the app.

ApplicationPolicy
Policy for an individual app.

ApplicationReport
Information reported about an installed app.

ApplicationReportingSettings
Settings controlling the behavior of application reports.

BlockAction
An action to block access to apps and data on a fully managed device or in a work profile. This action also triggers a device or work profile to displays a user-facing notification with information (where possible) on how to correct the compliance issue. Note: wipeAction must also be specified.

ChoosePrivateKeyRule
A rule for automatically choosing a private key and certificate to authenticate the device to a server.

Chunk

Command
A command.

ComplianceRule
A rule declaring which mitigating actions to take when a device is not compliant with its policy. For every rule, there is always an implicit mitigating action to set policy_compliant to false for the Device resource, and display a message on the device indicating that the device is not compliant with its policy. Other mitigating actions may optionally be taken as well, depending on the field values in the rule.

ContentRange
Implements the Content-Range header, for serialization only

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Device
A device owned by an enterprise. Unless otherwise noted, all fields are read-only and can't be modified by enterprises.devices.patch.

DeviceSettings
Information about security related device settings on device.

Display
Device display information.

DummyNetworkStream

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

EnrollmentToken
An enrollment token.

Enterprise
The configuration applied to an enterprise.

EnterpriseApplicationGetCall
Gets info about an application.

EnterpriseCreateCall
Creates an enterprise. This is the last step in the enterprise signup flow.

EnterpriseDeviceDeleteCall
Deletes a device. This operation wipes the device.

EnterpriseDeviceGetCall
Gets a device.

EnterpriseDeviceIssueCommandCall
Issues a command to a device. The Operation resource returned contains a Command in its metadata field. Use the get operation method to get the status of the command.

EnterpriseDeviceListCall
Lists devices for a given enterprise.

EnterpriseDeviceOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

EnterpriseDeviceOperationDeleteCall
Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

EnterpriseDeviceOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

EnterpriseDeviceOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED.NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users//operations. To override the binding, API services can add a binding such as "/v1/{name=users/}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

EnterpriseDevicePatchCall
Updates a device.

EnterpriseEnrollmentTokenCreateCall
Creates an enrollment token for a given enterprise.

EnterpriseEnrollmentTokenDeleteCall
Deletes an enrollment token. This operation invalidates the token, preventing its future use.

EnterpriseGetCall
Gets an enterprise.

EnterpriseMethods
A builder providing access to all methods supported on enterprise resources. It is not used directly, but through the AndroidManagement hub.

EnterprisePatchCall
Updates an enterprise.

EnterprisePolicyDeleteCall
Deletes a policy. This operation is only permitted if no devices are currently referencing the policy.

EnterprisePolicyGetCall
Gets a policy.

EnterprisePolicyListCall
Lists policies for a given enterprise.

EnterprisePolicyPatchCall
Updates or creates a policy.

EnterpriseWebAppCreateCall
Creates a web app.

EnterpriseWebAppDeleteCall
Deletes a web app.

EnterpriseWebAppGetCall
Gets a web app.

EnterpriseWebAppListCall
Lists web apps for a given enterprise.

EnterpriseWebAppPatchCall
Updates a web app.

EnterpriseWebTokenCreateCall
Creates a web token to access an embeddable managed Google Play web UI for a given enterprise.

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

ExternalData
Data hosted at an external location. The data is to be downloaded by Android Device Policy and verified against the hash.

HardwareInfo
Information about device hardware. The fields related to temperature thresholds are only available if hardwareStatusEnabled is true in the device's policy.

HardwareStatus
Hardware status. Temperatures may be compared to the temperature thresholds available in hardwareInfo to determine hardware health.

JsonServerError
A utility type which can decode a server response that indicates error

KeyedAppState
Keyed app state reported by the app.

LaunchAppAction
An action to launch an app.

ListDevicesResponse
Response to a request to list devices for a given enterprise.

ListOperationsResponse
The response message for Operations.ListOperations.

ListPoliciesResponse
Response to a request to list policies for a given enterprise.

ListWebAppsResponse
Response to a request to list web apps for a given enterprise.

ManagedConfigurationTemplate
The managed configurations template for the app, saved from the managed configurations iframe.

ManagedProperty
Managed property.

ManagedPropertyEntry
An entry of a managed property.

MemoryEvent
An event related to memory and storage measurements.

MemoryInfo
Information about device memory and storage.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

NetworkInfo
Device network info.

NonComplianceDetail
Provides detail about non-compliance with a policy setting.

NonComplianceDetailCondition
A compliance rule condition which is satisfied if there exists any matching NonComplianceDetail for the device. A NonComplianceDetail matches a NonComplianceDetailCondition if all the fields which are set within the NonComplianceDetailCondition match the corresponding NonComplianceDetail fields.

Operation
This resource represents a long-running operation that is the result of a network API call.

PackageNameList
A list of package names.

PasswordRequirements
Requirements for the password used to unlock a device.

PermissionGrant
Configuration for an Android permission and its grant state.

PersistentPreferredActivity
A default activity for handling intents that match a particular intent filter.

Policy
A policy resources represents a group settings that govern the behavior of a managed device and the apps installed on it.

PolicyEnforcementRule
A rule that defines the actions to take if a device or work profile is not compliant with the policy specified in settingName.

PowerManagementEvent
A power management event.

ProxyInfo
Configuration info for an HTTP proxy. For a direct proxy, set the host, port, and excluded_hosts fields. For a PAC script proxy, set the pac_uri field.

RangeResponseHeader

ResumableUploadHelper
A utility type to perform a resumable upload from start to end.

ServerError

ServerMessage

SetupAction
An action executed during setup.

SigninDetail
A resource containing sign in details for an enterprise.

SignupUrl
An enterprise signup URL.

SignupUrlCreateCall
Creates an enterprise signup URL.

SignupUrlMethods
A builder providing access to all methods supported on signupUrl resources. It is not used directly, but through the AndroidManagement hub.

SoftwareInfo
Information about device software.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC (https://github.com/grpc). Each Status message contains three pieces of data: error code, error message, and error details.You can find out more about this error model and how to work with it in the API Design Guide (https://cloud.google.com/apis/design/errors).

StatusReportingSettings
Settings controlling the behavior of status reports.

SystemUpdate
Configuration for managing system updates

TermsAndConditions
A terms and conditions page to be accepted during provisioning.

User
A user belonging to an enterprise.

UserFacingMessage
Provides a user-facing message with locale info. The maximum message length is 4096 characters.

WebApp
A web app.

WebAppIcon
An icon for a web app. Supported formats are: png, jpg and webp.

WebToken
A web token used to access the managed Google Play iframe.

WipeAction
An action to reset a fully managed device or delete a work profile. Note: blockAction must also be specified.

XUploadContentType
The X-Upload-Content-Type header.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

UnusedType
Identifies types which are not actually used by the API This might be a bug within the google API schema.

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.