Module google_androidmanagement1::api

Structs

AdvancedSecurityOverrides
Security policies set to the most secure values by default. To maintain the security posture of a device, we don't recommend overriding any of the default values.

AlwaysOnVpnPackage
Configuration for an always-on VPN connection.

AndroidManagement
Central instance to access all AndroidManagement related resource activities

ApiLevelCondition
A compliance rule condition which is satisfied if the Android Framework API level on the device doesn't meet a minimum requirement. There can only be one rule with this type of condition per policy.

AppTrackInfo
Id to name association of a app track.

Application
Information about an app.

ApplicationEvent
An app-related event.

ApplicationPermission
A permission required by the app.

ApplicationPolicy
Policy for an individual app.

ApplicationReport
Information reported about an installed app.

ApplicationReportingSettings
Settings controlling the behavior of application reports.

BlockAction
An action to block access to apps and data on a fully managed device or in a work profile. This action also triggers a device or work profile to displays a user-facing notification with information (where possible) on how to correct the compliance issue. Note: wipeAction must also be specified.

ChoosePrivateKeyRule
Controls apps' access to private keys. The rule determines which private key, if any, Android Device Policy grants to the specified app. Access is granted either when the app calls KeyChain.choosePrivateKeyAlias (https://developer.android.com/reference/android/security/KeyChain#choosePrivateKeyAlias%28android.app.Activity,%20android.security.KeyChainAliasCallback,%20java.lang.String[],%20java.security.Principal[],%20java.lang.String,%20int,%20java.lang.String%29) (or any overloads) to request a private key alias for a given URL, or for rules that are not URL-specific (that is, if urlPattern is not set, or set to the empty string or .*) on Android 11 and above, directly so that the app can call KeyChain.getPrivateKey (https://developer.android.com/reference/android/security/KeyChain#getPrivateKey%28android.content.Context,%20java.lang.String%29), without first having to call KeyChain.choosePrivateKeyAlias.When an app calls KeyChain.choosePrivateKeyAlias if more than one choosePrivateKeyRules matches, the last matching rule defines which key alias to return.

Command
A command.

CommonCriteriaModeInfo
Information about Common Criteria Mode—security standards defined in the Common Criteria for Information Technology Security Evaluation (https://www.commoncriteriaportal.org/) (CC).This information is only available if statusReportingSettings.commonCriteriaModeEnabled is true in the device's policy.

ComplianceRule
A rule declaring which mitigating actions to take when a device is not compliant with its policy. For every rule, there is always an implicit mitigating action to set policy_compliant to false for the Device resource, and display a message on the device indicating that the device is not compliant with its policy. Other mitigating actions may optionally be taken as well, depending on the field values in the rule.

ContactInfo
Contact details for LaForge enterprises.

ContentProviderEndpoint
This feature is not generally available.

Date
Represents a whole or partial calendar date, such as a birthday. The time of day and time zone are either specified elsewhere or are insignificant. The date is relative to the Gregorian Calendar. This can represent one of the following: A full date, with non-zero year, month, and day values A month and day value, with a zero year, such as an anniversary A year on its own, with zero month and day values A year and month value, with a zero day, such as a credit card expiration dateRelated types are google.type.TimeOfDay and google.protobuf.Timestamp.

Device
A device owned by an enterprise. Unless otherwise noted, all fields are read-only and can't be modified by enterprises.devices.patch.

DeviceSettings
Information about security related device settings on device.

Display
Device display information.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

EnrollmentToken
An enrollment token.

Enterprise
The configuration applied to an enterprise.

EnterpriseApplicationGetCall
Gets info about an application.

EnterpriseCreateCall
Creates an enterprise. This is the last step in the enterprise signup flow.

EnterpriseDeleteCall
This feature is not generally available yet. Deletes an enterprise.

EnterpriseDeviceDeleteCall
Deletes a device. This operation wipes the device.

EnterpriseDeviceGetCall
Gets a device.

EnterpriseDeviceIssueCommandCall
Issues a command to a device. The Operation resource returned contains a Command in its metadata field. Use the get operation method to get the status of the command.

EnterpriseDeviceListCall
Lists devices for a given enterprise.

EnterpriseDeviceOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

EnterpriseDeviceOperationDeleteCall
Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

EnterpriseDeviceOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

EnterpriseDeviceOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED.NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users//operations. To override the binding, API services can add a binding such as "/v1/{name=users/}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

EnterpriseDevicePatchCall
Updates a device.

EnterpriseEnrollmentTokenCreateCall
Creates an enrollment token for a given enterprise.

EnterpriseEnrollmentTokenDeleteCall
Deletes an enrollment token. This operation invalidates the token, preventing its future use.

EnterpriseGetCall
Gets an enterprise.

EnterpriseListCall
This feature is not generally available yet. Lists enterprises that are managed by an EMM. Only partial views are returned.

EnterpriseMethods
A builder providing access to all methods supported on enterprise resources. It is not used directly, but through the AndroidManagement hub.

EnterprisePatchCall
Updates an enterprise.

EnterprisePolicyDeleteCall
Deletes a policy. This operation is only permitted if no devices are currently referencing the policy.

EnterprisePolicyGetCall
Gets a policy.

EnterprisePolicyListCall
Lists policies for a given enterprise.

EnterprisePolicyPatchCall
Updates or creates a policy.

EnterpriseWebAppCreateCall
Creates a web app.

EnterpriseWebAppDeleteCall
Deletes a web app.

EnterpriseWebAppGetCall
Gets a web app.

EnterpriseWebAppListCall
Lists web apps for a given enterprise.

EnterpriseWebAppPatchCall
Updates a web app.

EnterpriseWebTokenCreateCall
Creates a web token to access an embeddable managed Google Play web UI for a given enterprise.

ExternalData
Data hosted at an external location. The data is to be downloaded by Android Device Policy and verified against the hash.

FreezePeriod
A system freeze period. When a device’s clock is within the freeze period, all incoming system updates (including security patches) are blocked and won’t be installed. When a device is outside the freeze period, normal update behavior applies. Leap years are ignored in freeze period calculations, in particular: * If Feb. 29th is set as the start or end date of a freeze period, the freeze period will start or end on Feb. 28th instead. * When a device’s system clock reads Feb. 29th, it’s treated as Feb. 28th. * When calculating the number of days in a freeze period or the time between two freeze periods, Feb. 29th is ignored and not counted as a day.

HardwareInfo
Information about device hardware. The fields related to temperature thresholds are only available if hardwareStatusEnabled is true in the device's policy.

HardwareStatus
Hardware status. Temperatures may be compared to the temperature thresholds available in hardwareInfo to determine hardware health.

KeyedAppState
Keyed app state reported by the app.

KioskCustomization
Settings controlling the behavior of a device in kiosk mode. To enable kiosk mode, set kioskCustomLauncherEnabled to true or specify an app in the policy with installType KIOSK.

LaunchAppAction
An action to launch an app.

ListDevicesResponse
Response to a request to list devices for a given enterprise.

ListEnterprisesResponse
This feature is not generally available yet. Response to a request to list enterprises.

ListOperationsResponse
The response message for Operations.ListOperations.

ListPoliciesResponse
Response to a request to list policies for a given enterprise.

ListWebAppsResponse
Response to a request to list web apps for a given enterprise.

ManagedConfigurationTemplate
The managed configurations template for the app, saved from the managed configurations iframe.

ManagedProperty
Managed property.

ManagedPropertyEntry
An entry of a managed property.

MemoryEvent
An event related to memory and storage measurements.

MemoryInfo
Information about device memory and storage.

NetworkInfo
Device network info.

NonComplianceDetail
Provides detail about non-compliance with a policy setting.

NonComplianceDetailCondition
A compliance rule condition which is satisfied if there exists any matching NonComplianceDetail for the device. A NonComplianceDetail matches a NonComplianceDetailCondition if all the fields which are set within the NonComplianceDetailCondition match the corresponding NonComplianceDetail fields.

OncCertificateProvider
This feature is not generally available.

Operation
This resource represents a long-running operation that is the result of a network API call.

PackageNameList
A list of package names.

PasswordRequirements
Requirements for the password used to unlock a device.

PermissionGrant
Configuration for an Android permission and its grant state.

PersistentPreferredActivity
A default activity for handling intents that match a particular intent filter. Note: To set up a kiosk, use InstallType to KIOSK rather than use persistent preferred activities.

PersonalApplicationPolicy
Policies for apps in the personal profile of a company-owned device with a work profile.

PersonalUsagePolicies
Policies controlling personal usage on a company-owned device with a work profile.

Policy
A policy resource represents a group of settings that govern the behavior of a managed device and the apps installed on it.

PolicyEnforcementRule
A rule that defines the actions to take if a device or work profile is not compliant with the policy specified in settingName.

PostureDetail
Additional details regarding the security posture of the device.

PowerManagementEvent
A power management event.

ProxyInfo
Configuration info for an HTTP proxy. For a direct proxy, set the host, port, and excluded_hosts fields. For a PAC script proxy, set the pac_uri field.

SecurityPosture
The security posture of the device, as determined by the current device state and the policies applied.

SetupAction
An action executed during setup.

SigninDetail
A resource containing sign in details for an enterprise.

SignupUrl
An enterprise signup URL.

SignupUrlCreateCall
Creates an enterprise signup URL.

SignupUrlMethods
A builder providing access to all methods supported on signupUrl resources. It is not used directly, but through the AndroidManagement hub.

SoftwareInfo
Information about device software.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC (https://github.com/grpc). Each Status message contains three pieces of data: error code, error message, and error details.You can find out more about this error model and how to work with it in the API Design Guide (https://cloud.google.com/apis/design/errors).

StatusReportingSettings
Settings controlling the behavior of status reports.

SystemUpdate
Configuration for managing system updates

SystemUpdateInfo
Information about a potential pending system update.

TelephonyInfo
Telephony information associated with a given SIM card on the device. Only supported on fully managed devices starting from Android API level 23.

TermsAndConditions
A terms and conditions page to be accepted during provisioning.

User
A user belonging to an enterprise.

UserFacingMessage
Provides a user-facing message with locale info. The maximum message length is 4096 characters.

WebApp
A web app.

WebAppIcon
An icon for a web app. Supported formats are: png, jpg and webp.

WebToken
A web token used to access the managed Google Play iframe.

WipeAction
An action to reset a fully managed device or delete a work profile. Note: blockAction must also be specified.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.