Struct google_androidmanagement1::EnterpriseMethods

pub struct EnterpriseMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on enterprise resources. It is not used directly, but through the AndroidManagement hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_androidmanagement1 as androidmanagement1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use androidmanagement1::AndroidManagement;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = AndroidManagement::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `applications_get(...)`, `create(...)`, `devices_delete(...)`, `devices_get(...)`, `devices_issue_command(...)`, `devices_list(...)`, `devices_operations_cancel(...)`, `devices_operations_delete(...)`, `devices_operations_get(...)`, `devices_operations_list(...)`, `devices_patch(...)`, `enrollment_tokens_create(...)`, `enrollment_tokens_delete(...)`, `get(...)`, `patch(...)`, `policies_delete(...)`, `policies_get(...)`, `policies_list(...)`, `policies_patch(...)`, `web_apps_create(...)`, `web_apps_delete(...)`, `web_apps_get(...)`, `web_apps_list(...)`, `web_apps_patch(...)` and `web_tokens_create(...)`
// to build up your call.
let rb = hub.enterprises();

Methods

impl<'a, C, A> EnterpriseMethods<'a, C, A>

pub fn policies_list(&self, parent: &str) -> EnterprisePolicyListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists policies for a given enterprise.

Arguments

parent - The name of the enterprise in the form enterprises/{enterpriseId}.

pub fn web_apps_delete(
&self,
name: &str
) -> EnterpriseWebAppDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a web app.

Arguments

name - The name of the web app in the form enterprises/{enterpriseId}/webApps/{packageName}.

pub fn policies_patch(
&self,
request: Policy,
name: &str
) -> EnterprisePolicyPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates or creates a policy.

Arguments

request - No description provided.

name - The name of the policy in the form enterprises/{enterpriseId}/policies/{policyId}.

pub fn create(&self, request: Enterprise) -> EnterpriseCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an enterprise. This is the last step in the enterprise signup flow.

Arguments

request - No description provided.

pub fn web_apps_create(
&self,
request: WebApp,
parent: &str
) -> EnterpriseWebAppCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a web app.

Arguments

request - No description provided.

parent - The name of the enterprise in the form enterprises/{enterpriseId}.

pub fn enrollment_tokens_create(
&self,
request: EnrollmentToken,
parent: &str
) -> EnterpriseEnrollmentTokenCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates an enrollment token for a given enterprise.

Arguments

request - No description provided.

parent - The name of the enterprise in the form enterprises/{enterpriseId}.

pub fn web_apps_list(&self, parent: &str) -> EnterpriseWebAppListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists web apps for a given enterprise.

Arguments

parent - The name of the enterprise in the form enterprises/{enterpriseId}.

pub fn devices_list(&self, parent: &str) -> EnterpriseDeviceListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists devices for a given enterprise.

Arguments

parent - The name of the enterprise in the form enterprises/{enterpriseId}.

pub fn policies_delete(
&self,
name: &str
) -> EnterprisePolicyDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a policy. This operation is only permitted if no devices are currently referencing the policy.

Arguments

name - The name of the policy in the form enterprises/{enterpriseId}/policies/{policyId}.

pub fn devices_operations_cancel(
&self,
name: &str
) -> EnterpriseDeviceOperationCancelCall<'a, C, A>

Create a builder to help you perform the following task:

Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

Arguments

name - The name of the operation resource to be cancelled.

pub fn devices_operations_delete(
&self,
name: &str
) -> EnterpriseDeviceOperationDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

Arguments

name - The name of the operation resource to be deleted.

pub fn devices_operations_get(
&self,
name: &str
) -> EnterpriseDeviceOperationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn devices_operations_list(
&self,
name: &str
) -> EnterpriseDeviceOperationListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED.NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users//operations. To override the binding, API services can add a binding such as "/v1/{name=users/}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Arguments

name - The name of the operation's parent resource.

pub fn devices_patch(
&self,
request: Device,
name: &str
) -> EnterpriseDevicePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a device.

Arguments

request - No description provided.

name - The name of the device in the form enterprises/{enterpriseId}/devices/{deviceId}.

pub fn patch(
&self,
request: Enterprise,
name: &str
) -> EnterprisePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an enterprise.

Arguments

request - No description provided.

name - The name of the enterprise in the form enterprises/{enterpriseId}.

pub fn policies_get(&self, name: &str) -> EnterprisePolicyGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a policy.

Arguments

name - The name of the policy in the form enterprises/{enterpriseId}/policies/{policyId}.

pub fn web_tokens_create(
&self,
request: WebToken,
parent: &str
) -> EnterpriseWebTokenCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a web token to access an embeddable managed Google Play web UI for a given enterprise.

Arguments

request - No description provided.

parent - The name of the enterprise in the form enterprises/{enterpriseId}.

pub fn web_apps_patch(
&self,
request: WebApp,
name: &str
) -> EnterpriseWebAppPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a web app.

Arguments

request - No description provided.

name - The name of the web app in the form enterprises/{enterpriseId}/webApps/{packageName}.

pub fn enrollment_tokens_delete(
&self,
name: &str
) -> EnterpriseEnrollmentTokenDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes an enrollment token. This operation invalidates the token, preventing its future use.

Arguments

name - The name of the enrollment token in the form enterprises/{enterpriseId}/enrollmentTokens/{enrollmentTokenId}.

pub fn devices_issue_command(
&self,
request: Command,
name: &str
) -> EnterpriseDeviceIssueCommandCall<'a, C, A>

Create a builder to help you perform the following task:

Issues a command to a device. The Operation resource returned contains a Command in its metadata field. Use the get operation method to get the status of the command.

Arguments

request - No description provided.

name - The name of the device in the form enterprises/{enterpriseId}/devices/{deviceId}.

pub fn devices_delete(&self, name: &str) -> EnterpriseDeviceDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a device. This operation wipes the device.

Arguments

name - The name of the device in the form enterprises/{enterpriseId}/devices/{deviceId}.

pub fn web_apps_get(&self, name: &str) -> EnterpriseWebAppGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a web app.

Arguments

name - The name of the web app in the form enterprises/{enterpriseId}/webApp/{packageName}.

pub fn devices_get(&self, name: &str) -> EnterpriseDeviceGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a device.

Arguments

name - The name of the device in the form enterprises/{enterpriseId}/devices/{deviceId}.

pub fn get(&self, name: &str) -> EnterpriseGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets an enterprise.

Arguments

name - The name of the enterprise in the form enterprises/{enterpriseId}.

pub fn applications_get(
&self,
name: &str
) -> EnterpriseApplicationGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets info about an application.

Arguments

name - The name of the application in the form enterprises/{enterpriseId}/applications/{package_name}.